refactor: flytta ExpandableCodeBlock till CodeBlock och förenkla import
