+++ src/components/common/CodeBlock.tsx
@@ -7,6 +7,12 @@ interface CodeBlockProps {
   defaultExpanded?: boolean
 }
 
+interface ExpandableCodeBlockProps {
+  filename: string
+  code: string
+  defaultExpanded?: boolean
+}
+
 export function CodeBlock({ children, title, defaultExpanded = false }: CodeBlockProps) {
   const [isExpanded, setIsExpanded] = useState(defaultExpanded)
 
@@ -50,6 +56,31 @@ export function CodeBlock({ children, title, defaultExpanded = false }: CodeBloc
         <div className="p-3 text-xs font-mono">
           <pre>{children}</pre>
         </div>
+      )}
+    </div>
+  )
+}
+
+export function ExpandableCodeBlock({ filename, code, defaultExpanded = false }: ExpandableCodeBlockProps) {
+  const [isExpanded, setIsExpanded] = useState(defaultExpanded)
+
+  return (
+    <div className="my-4">
+      <button
+        onClick={() => setIsExpanded(!isExpanded)}
+        className="w-full flex items-center gap-2 px-4 py-2 bg-[#2D6A4F] text-white text-sm font-mono rounded-t-lg border-b border-white/10 hover:bg-[#2D6A4F]/80 transition-colors"
+      >
+        <span className="flex-1 text-left font-medium">{filename}</span>
+        {isExpanded ? (
+          <ChevronDown className="w-4 h-4" />
+        ) : (
+          <ChevronRight className="w-4 h-4" />
+        )}
+      </button>
+      {isExpanded && (
+        <pre className="hljs bg-[#0d1117] border border-white/10 text-sm overflow-x-auto rounded-b-lg m-0 p-4">
+          <code dangerouslySetInnerHTML={{ __html: code }} />
+        </pre>
       )}
     </div>
   )
